atelet: Smush cred broker and worker reporting into "AteomSupportService"
